requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
aspose-cad==24.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
cement==2.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.7
dearpygui==1.11.1
Deprecated==1.2.14
dnspython==2.6.1
docstring_parser==0.16
easydict==1.13
einops==0.7.0
email_validator==2.2.0
embreex==2.17.7.post5
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.2.0
gdown==5.1.0
gif==23.3.0
glcontext==2.5.0
gradio==4.8.0
gradio_client==0.7.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
imgviz==1.7.5
importlib_resources==6.4.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiui @ git+https://github.com/ashawkey/kiuikit@ebf6dd070b936561b325d9e51044ed18a3241268
kiwi==3.0.3
kiwisolver==1.4.5
kornia_rs==0.1.3
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.2.2
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
mdurl==0.1.2
moderngl==5.10.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast/@c5caf7bdb8a2448acc491a9faa47753972edd380
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
objprint==0.2.3
omegaconf==2.3.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pathspec==0.10.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
psutil==6.0.0
PyAutoGUI==0.9.54
pycocotools==2.0.8
pycollada==0.8
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyGetWindow==0.0.9
pygltflib==1.16.2
Pygments==2.18.0
pyliblzfse==0.4.1
PyMatting==1.1.12
PyMCubes==0.1.6
pymeshlab==2023.12.post1
PyMsgBox==1.0.9
PyOpenGL==3.1.7
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
rembg==2.0.57
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
Rtree==1.3.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.1
scipy==1.13.0
selenium==4.23.1
semantic-version==2.8.5
shapely==2.0.5
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
svg.path==6.3
sympy==1.12
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2024.4.24
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchmcubes @ git+https://github.com/tatsy/torchmcubes.git@cbb3c3795b1e168bf81e8dee28623eaf5c33cd1c
torchvision==0.18.0+cu118
tqdm==4.66.4
transformers==4.35.0
trimesh==4.0.5
trio==0.26.0
trio-websocket==0.11.1
triton==2.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
varname==0.13.1
viser==0.2.3
watchfiles==0.22.0
wcwidth==0.1.4
websocket-client==1.8.0
websockets==11.0.3
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
xatlas==0.0.9
xxhash==3.4.1
yourdfpy==0.0.56
